refactored FlightDeck.Editor into PackageController
