+++ apps/jetpack/templates/edit.html
@@ -8,10 +8,8 @@
 	<link rel="stylesheet" href="/media/jetpack/css/FlightDeck.Autocomplete.css" type="text/css" media="screen" />
 	<link rel="stylesheet" href="/media/jetpack/css/tree.css" type="text/css" media="screen" />
 	<link rel="stylesheet" href="/media/jetpack/css/collapse.css" type="text/css" media="screen" />
-    <script src="/media/lib/ace/ace-uncompressed.js"></script>
 	<script src="/media/lib/tree.js"></script>
 
-	<script src="/media/jetpack/js/FlightDeck.Editor.js"></script>
 	<script src="/media/lib/ace/ace.js"></script>
 	<script src="/media/lib/ace/mode-javascript.js"></script>
 	<script src="/media/lib/ace/mode-css.js"></script>
