+++ apps/jetpack/templates/view.html
@@ -16,7 +16,6 @@
     <script src="/media/jetpack/js/editor-min.js"></script>
     {% endif %}
 
-	<script src="/media/jetpack/js/FlightDeck.Editor.js"></script>
 	<script src="/media/lib/ace/mode-javascript.js"></script>
 	<script src="/media/lib/ace/mode-css.js"></script>
 	<script src="/media/lib/ace/mode-html.js"></script>
